requirements.txt
streamlit>=1.32
pypdf>=4.0.0
Pillow>=10.0.0
